requirements.txt
xmlschema>=2.0.0
pandas>=1.5.0
tqdm>=4.64.0
openpyxl>=3.0.0
pytest>=7.0.0
mypy>=1.0.0

# Streamlit Dependencies
streamlit>=1.28.0
plotly>=5.17.0
psycopg2-binary>=2.9.5
seaborn>=0.12.0
matplotlib>=3.7.0
numpy>=1.24.0
